requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
distlib==0.3.8
Django==5.1.3
djangorestframework==3.15.2
filelock==3.14.0
platformdirs==4.2.1
soupsieve==2.5
sqlparse==0.5.1
virtualenv==20.26.1
